# client-api/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Matrix Client-Server Api",
    "version" : "r0.3.0",
    "title" : "Matrix Client-Server Api",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Anatoly",
      "url" : "https://github.com/ma1uta/jeon",
      "email" : "sablintolya@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "matrix.org:8448",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/_matrix/client/r0/login/cas/redirect" : {
      "get" : {
        "tags" : [ "CAS" ],
        "summary" : "A web-based Matrix client should instruct the user's browser to navigate to this endpoint in order to log in via CAS.",
        "description" : "The server MUST respond with an HTTP redirect to the CAS interface. The URI MUST include a service parameter giving the path of the /login/cas/ticket endpoint (including the redirectUrl query parameter).",
        "operationId" : "redirect",
        "parameters" : [ {
          "name" : "redirectUrl",
          "in" : "query",
          "description" : "URI to which the user will be redirected after the homeserver has authenticated the user with CAS.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/EmptyResponse"
            }
          },
          "302" : {
            "description" : "A redirect to the CAS interface"
          }
        }
      }
    },
...
    "/_matrix/client/r0/register/email/requestToken" : {
      "post" : {
        "tags" : [ "Account" ],
        "summary" : "Request token.",
        "description" : "Proxies the identity server API validate/email/requestToken, but first checks that the given email address is not already associated with an account on this Home Server. Note that, for consistency, this API takes JSON objects, though the Identity Server API takes x-www-form-urlencoded parameters. See the Identity Server API for further information.",
        "operationId" : "requestToken",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "request",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RequestToken"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "An email has been sent to the specified address.",
            "schema" : {
              "$ref" : "#/definitions/EmptyResponse"
            }
          },
          "400" : {
            "description" : "Part of the request was invalid."
          }
        }
      }
    },
...
    "/_matrix/client/r0/rooms/{roomId}/context/{eventId}" : {
      "get" : {
        "tags" : [ "EventContext" ],
        "summary" : "This API returns a number of events that happened just before and after the specified event.",
        "description" : "This allows clients to get the context surrounding an event.",
        "operationId" : "context",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "description" : "The room to get events from.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "eventId",
          "in" : "path",
          "description" : "The event to get context around.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The maximum number of events to return. Default: 10.",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "The events and state surrounding the requested event.",
            "schema" : {
              "$ref" : "#/definitions/EventContextResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "EmptyResponse" : {
      "type" : "object",
      "description" : "Empty response"
    },
...
    "Event" : {
      "type" : "object",
      "required" : [ "event_id", "origin_server_ts", "room_id", "sender", "state_key", "type" ],
      "properties" : {
        "content" : {
          "description" : "The fields in this object will vary depending on the type of event. When interacting with the REST API, this is the HTTP body.",
          "$ref" : "#/definitions/EventContent"
        },
        "type" : {
          "type" : "string",
          "description" : "The type of event. This SHOULD be namespaced similar to Java package naming conventions."
        },
        "sender" : {
          "type" : "string",
          "description" : "Contains the fully-qualified ID of the user who sent this event."
        },
        "unsigned" : {
          "description" : "Contains optional extra information about the event.",
          "$ref" : "#/definitions/Unsigned"
        },
        "event_id" : {
          "type" : "string",
          "description" : "The globally unique event identifier."
        },
        "room_id" : {
          "type" : "string",
          "description" : "The ID of the room associated with this event."
        },
        "origin_server_ts" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Timestamp in milliseconds on originating homeserver when this event was sent."
        },
        "prev_content" : {
          "description" : "The previous content for this event. If there is no previous content, this key will be missing.",
          "$ref" : "#/definitions/EventContent"
        },
        "state_key" : {
          "type" : "string",
          "description" : " A unique key which defines the overwriting semantics for this piece of room state. This value is often a zero-length string. The presence of this key makes this event a State Event. The key MUST NOT start with '_'."
        },
        "invite_room_state" : {
          "type" : "array",
          "description" : "This contains an array of StrippedState Events. These events provide information on a subset of state events such as the room name.",
          "items" : {
            "$ref" : "#/definitions/StrippedState"
          }
        }
      },
      "description" : "Event."
    },
    "EventContent" : {
      "type" : "object",
      "discriminator" : "type",
      "description" : "Parent class of all event contents part in the Event."
    },
...
    "EventContextResponse" : {
      "type" : "object",
      "properties" : {
        "start" : {
          "type" : "string",
          "description" : "Pagination token for the start of the chunk."
        },
        "end" : {
          "type" : "string",
          "description" : "Pagination token for the end of the chunk."
        },
        "profile_info" : {
          "type" : "object",
          "description" : "The historic profile information of the users that sent the events returned.",
          "additionalProperties" : {
            "$ref" : "#/definitions/Profile"
          }
        },
        "events_before" : {
          "type" : "array",
          "description" : "Events just before the result.",
          "items" : {
            "$ref" : "#/definitions/Event"
          }
        },
        "events_after" : {
          "type" : "array",
          "description" : "Events just after the result.",
          "items" : {
            "$ref" : "#/definitions/Event"
          }
        }
      },
      "description" : "Event context."
    },
...
    "Profile" : {
      "type" : "object",
      "properties" : {
        "displayname" : {
          "type" : "string",
          "description" : "The user's display name if they have set one, otherwise not present."
        },
        "avatar_url" : {
          "type" : "string",
          "description" : "The user's avatar URL if they have set one, otherwise not present."
        }
      },
      "description" : "JSON body for profile api (profile)."
    },
...
    "RequestToken" : {
      "type" : "object",
      "required" : [ "client_secret", "email", "send_attempt" ],
      "properties" : {
        "email" : {
          "type" : "string",
          "description" : "The email address"
        },
        "id_server" : {
          "type" : "string",
          "description" : "The ID server to send the onward request to as a hostname with an appended colon and port number if the port is not the default."
        },
        "client_secret" : {
          "type" : "string",
          "description" : "Client-generated secret string used to protect this session."
        },
        "send_attempt" : {
          "type" : "string",
          "description" : "Used to distinguish protocol level retries from requests to re-send the email."
        }
      },
      "description" : "Request for the proxies the identity server API validate/email/requestToken."
    },
...
    "StrippedState" : {
      "type" : "object",
      "required" : [ "content", "state_key", "type" ],
      "properties" : {
        "content" : {
          "type" : "object",
          "description" : "The content for the event.",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "type" : {
          "type" : "string",
          "description" : "The type for the event."
        },
        "state_key" : {
          "type" : "string",
          "description" : "The state_key for the event."
        }
      },
      "description" : "Provides information on a subset of state events such as the room name."
    },
...
    "Unsigned" : {
      "type" : "object",
      "properties" : {
        "age" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "The time in milliseconds that has elapsed since the event was sent. This field is generated by the local homeserver, and may be incorrect if the local time on at least one of the two servers is out of sync, which can cause the age to either be negative or greater than it actually is."
        },
        "redacted_because" : {
          "description" : "The event that redacted this event, if any.",
          "$ref" : "#/definitions/Event"
        },
        "transaction_id" : {
          "type" : "string",
          "description" : "The client-supplied transaction ID, if the client being given the event is the same one which sent it."
        },
        "prev_content" : {
          "type" : "object",
          "description" : "The previous content for this state. This will be present only for state events appearing in the timeline. If this is not a state event, or there is no previous content, this key will be missing.",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      },
      "description" : "Unsigned addition data."
    },
...
  }
}